requirements.txt
python>=3.9
cudatoolkit
matplotlib
numpy
pandas
plotly
pyarrow
scipy
scikit-learn
seaborn
snakemake-minimal
tqdm
ujson
jupyter
ipykernel
ipywidgets
nb_conda_kernels
stanza
umap-learn
huggingface::transformers
pytorch::pytorch
